slight simplifications
